django-aiproviders/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "django-aiproviders"
version = "0.1.0"
description = "Django integration for AI provider management. Provides Django models and admin for the aiproviders library."
readme = "README.md"
authors = [{ name = "Octolo", email = "dev@octolo.tech" }]
license = { text = "MIT" }
requires-python = ">=3.10"
keywords = [
  "django",
  "ai",
  "llm",
  "openai",
  "provider",
  "django-models",
  "python",
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "Framework :: Django",
  "Framework :: Django :: 3.2",
  "Framework :: Django :: 4.0",
  "Framework :: Django :: 4.1",
  "Framework :: Django :: 4.2",
  "Framework :: Django :: 5.0",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "Django>=3.2",
    "aiproviders>=0.1.0",
    "django-boosted>=0.1.0",
    "django-providerkit>=0.1.0",
    "django-virtualqueryset>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/hicinformatic/python-aiproviders"
Repository = "https://github.com/hicinformatic/python-aiproviders"
Documentation = "https://github.com/hicinformatic/python-aiproviders/tree/main/django-aiproviders#readme"
Issues = "https://github.com/hicinformatic/python-aiproviders/issues"

[tool.setuptools]
license-files = ["../LICENSE"]

[tool.setuptools.packages.find]
where = ["src"]
include = ["django_aiproviders*"]
exclude = ["tests*"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "tests.settings"
testpaths = ["tests"]
pythonpath = ["src"]
